backend/requirements.txt
fastapi
uvicorn
pydantic
fastapi-pagination
asyncpg
SQLAlchemy
python-dotenv
alembic
pydantic-settings